requirements_macOS.txt
python-dotenv
yfinance
openpyxl
openai-agents
asyncio 
nest_asyncio
transformers
scipy
tensorflow
torch
tf-keras
markdown-pdf